+++ src/legacy/picker/modules/picker.module.ts
@@ -36,7 +36,9 @@ import {LoopScrollDataDirective, LoopScrollDirective} from '../directives';
     ],
     exports:
     [
-        DateTimePickerLegacyComponent
+        DateTimePickerLegacyComponent,
+        LoopScrollDataDirective,
+        LoopScrollDirective,
     ]
 })
 export class DateTimeLegacyPickerModule