+++ src/modules/dateTimePicker/components/index.ts
@@ -3,4 +3,6 @@ export * from './dateTimePicker/dateTimePicker.component';
 export * from './dateTimePicker/dateTimePicker.interface';
 export * from './dayPicker/dayPicker.component';
 export * from './monthPicker/monthPicker.component';
+export * from './rollerTimePicker/rollerTimePicker.component';
+export * from './rollerTimePicker/rollerTimePicker.data';
 export * from './yearPicker/yearPicker.component';